Move Message to its own file.
